src/requirements.txt
flask >= 2.0.2
jsonify >= 0.5
Flask-SQLAlchemy >= 2.5.1
Flask-RESTful >= 0.3.9
Flask-Migrate >= 3.1.0
python-dotenv >= 0.19.2
marshmallow >= 3.14.1